cluster-ui: fix sort on tables
Sort on tables of Statement, Transaction and Sessions page were
not using the right value from the url regarding sorting. It was
using the right sortKey, but always doing ascending. Fix done so
it uses the right value for the boolean.

Since the sortKet might be different for different users on Statement
Page (difference on columns being displayed), use the name of the
columns, instead of the key.

Created based on PR https://github.com/cockroachdb/ui/pull/314
Changes to be added to version 0.2.15

Release note (bug fix): Sort on tables are now picking up the
correct value from the url.